refactor(runtime): move runtime code
