+++ packages/rust-runtime/js/process.ts
@@ -1,0 +1,5 @@
+(globalThis => {
+  globalThis.process = {
+    env: {},
+  };
+})(globalThis);
